rename react components
