+++ src/components/Panel.tsx
@@ -1,7 +1,7 @@
 import { useStore } from "@nanostores/react";
 import { $feature, $showInfoPanel, hideInfoPanel } from "../store/feature";
 import "./panel.css";
-import { OpeningHours } from "./OpeningHours";
+import { OpeningHours } from "./opening-hours";
 import { formatAddress } from "../utils/format-address";
 import KindBadge from "./kind-badge";
