+++ src/components/opening-hours.tsx
