+++ src/pages/index.astro
@@ -1,6 +1,6 @@
 ---
-import { Map } from "../components/Map";
-import { Panel } from "../components/Panel";
+import { Map } from "../components/map";
+import { Panel } from "../components/panel";
 import Layout from "../layouts/Layout.astro";
 ---
